Avoiding cross-platform pitfalls (and eliminating the crash described in the bug #21)

+++ avxsynth/builtinfunctions/src/filters/turnfunc.cpp
@@ -35,6 +35,7 @@
 // import and export plugins, or graphical user interfaces.
 
 #include "stdafx.h"
+#include <inttypes.h>
 
 /*
 ** Turn. version 0.1
@@ -93,8 +94,8 @@ void TurnRGB32(const unsigned char *srcp, unsigned char *dstp, const int rowsize
 			   const int height, const int src_pitch, const int dst_pitch,
 			   const int direction)
 {
-	unsigned long *l_srcp = (unsigned long *)srcp;
-	unsigned long *l_dstp = (unsigned long *)dstp;
+	uint32_t *l_srcp = (uint32_t *)srcp;
+	uint32_t *l_dstp = (uint32_t *)dstp;
 	int l_rowsize = rowsize/4;
 	int l_src_pitch = src_pitch/4;
 	int l_dst_pitch = dst_pitch/4;
